Disable R8 for now

+++ build-ci.sh
@@ -15,4 +15,6 @@ mkdir -p $MAPPINGS_OUT_DIR
 bash ./gradlew $BUILD_COMMAND
 
 mv build/outputs/apk/*/*/*.apk $APKS_OUT_DIR/Lawnchair.apk
-mv build/outputs/mapping/*/*/mapping.txt $MAPPINGS_OUT_DIR/mapping.txt
+if [ -f build/outputs/mapping/*/*/mapping.txt ]; then
+    mv build/outputs/mapping/*/*/mapping.txt $MAPPINGS_OUT_DIR/mapping.txt
+fi
